requirements.txt
bcrypt==4.0.1
blinker==1.5
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
ioe-email-stuff==0.0.11
itsdangerous==2.0.1
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
MarkupSafe==2.1.1
more-itertools==9.0.0
Pillow==9.4.0
pywin32-ctypes==0.2.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
